added Earthquakecontent
