+++ pubspec.yaml
@@ -70,6 +70,7 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - img/
+    - img/SPR/
